+++ storage/include/Storage.h
@@ -17,9 +17,9 @@ namespace minidb::storage {
 
 struct IndexFileMetaData {
     char          tag[ 13 ] = "Minidb Index";
-    std::uint32_t table_num;
-    std::uint32_t table_define_begin;
-    std::uint32_t table_define_end;
+    std::uint64_t table_num;
+    std::uint64_t table_define_begin;
+    std::uint64_t table_define_end;
 };
 
 struct DataFileMetaData {
@@ -27,12 +27,12 @@ struct DataFileMetaData {
 };
 
 struct TableDefineMetaData {
-    std::uint32_t table_define_length;
-    std::uint32_t table_root_addr;
-    std::uint32_t first_leaf_addr;
-    std::uint32_t last_leaf_addr;
-    std::uint32_t column_num;
-    std::uint32_t table_name_length;
+    std::uint64_t table_define_length;
+    std::uint64_t table_root_addr;
+    std::uint64_t first_leaf_addr;
+    std::uint64_t last_leaf_addr;
+    std::uint64_t column_num;
+    std::uint64_t table_name_length;
 };
 
 class Storage {
@@ -44,28 +44,28 @@ public:
     std::unordered_map<std::string, TableInfo> table_info_map;
 
     bool       scan_tables();
-    TableInfo  scan_table(std::uint32_t &current_addr);
-    ColumnInfo scan_column(std::uint32_t &current_addr);
+    TableInfo  scan_table(std::uint64_t &current_addr);
+    ColumnInfo scan_column(std::uint64_t &current_addr);
 
     bool       new_table(const ast::SQLCreateTableStatement &creat_statement);
     ColumnInfo write_column_define(const ast::SQLColumnDefine &column_define,
-                                   std::uint32_t &             current_addr);
+                                   std::uint64_t &             current_addr);
 
-    bool insert_data(const std::string &table_name, std::int32_t key,
+    bool insert_data(const std::string &table_name, std::int64_t key,
                      const SQLBinaryData &data);
 
-    SQLBinaryData search_data(const std::string &table_name, std::int32_t key);
+    SQLBinaryData search_data(const std::string &table_name, std::int64_t key);
 
     /**
      * @brief 写入指定大小的二进制数据，同时使current_addr往后移动
      */
-    bool write_binary(const void *data, std::uint32_t size,
-                      std::uint32_t &current_addr);
+    bool write_binary(const void *data, std::uint64_t size,
+                      std::uint64_t &current_addr);
 
-    std::uint32_t              table_num;
-    static const std::uint32_t pageSize           = 4096;
-    static const std::uint32_t tableNumAddr       = 7;
-    std::uint32_t              table_define_begin = 19;
-    std::uint32_t              table_define_end   = 19;
+    std::uint64_t              table_num;
+    static const std::uint64_t pageSize           = 4096;
+    static const std::uint64_t tableNumAddr       = 7;
+    std::uint64_t              table_define_begin = 19;
+    std::uint64_t              table_define_end   = 19;
 };
 } // namespace minidb::storage
